MNT split smote file into a submodule
