+++ imblearn/over_sampling/_smote/__init__.py
@@ -1,0 +1,17 @@
+from .base import SMOTE
+from .base import SMOTEN
+from .base import SMOTENC
+
+from .cluster import KMeansSMOTE
+
+from .filter import BorderlineSMOTE
+from .filter import SVMSMOTE
+
+__all__ = [
+    "SMOTE",
+    "SMOTEN",
+    "SMOTENC",
+    "KMeansSMOTE",
+    "BorderlineSMOTE",
+    "SVMSMOTE",
+]
