+++ imblearn/over_sampling/_smote/filter.py
@@ -1,0 +1,433 @@
+﻿"""SMOTE variant applying some filtering before the generation process."""
+
+# Authors: Guillaume Lemaitre <g.lemaitre58@gmail.com>
+#          Fernando Nogueira
+#          Christos Aridas
+#          Dzianis Dudnik
+# License: MIT
+
+import numpy as np
+from scipy import sparse
+
+from sklearn.base import clone
+from sklearn.svm import SVC
+from sklearn.utils import check_random_state
+from sklearn.utils import _safe_indexing
+
+from ..base import BaseOverSampler
+from ...exceptions import raise_isinstance_error
+from ...utils import check_neighbors_object
+from ...utils import Substitution
+from ...utils._docstring import _n_jobs_docstring
+from ...utils._docstring import _random_state_docstring
+from ...utils._validation import _deprecate_positional_args
+
+from .base import BaseSMOTE
+
+
+@Substitution(
+    sampling_strategy=BaseOverSampler._sampling_strategy_docstring,
+    n_jobs=_n_jobs_docstring,
+    random_state=_random_state_docstring,
+)
+class BorderlineSMOTE(BaseSMOTE):
+    """Over-sampling using Borderline SMOTE.
+
+    This algorithm is a variant of the original SMOTE algorithm proposed in
+    [2]_. Borderline samples will be detected and used to generate new
+    synthetic samples.
+
+    Read more in the :ref:`User Guide <smote_adasyn>`.
+
+    Parameters
+    ----------
+    {sampling_strategy}
+
+    {random_state}
+
+    k_neighbors : int or object, default=5
+        If ``int``, number of nearest neighbours to used to construct synthetic
+        samples.  If object, an estimator that inherits from
+        :class:`~sklearn.neighbors.base.KNeighborsMixin` that will be used to
+        find the k_neighbors.
+
+    {n_jobs}
+
+    m_neighbors : int or object, default=10
+        If int, number of nearest neighbours to use to determine if a minority
+        sample is in danger. If object, an estimator that inherits
+        from :class:`~sklearn.neighbors.base.KNeighborsMixin` that will be used
+        to find the m_neighbors.
+
+    kind : {{"borderline-1", "borderline-2"}}, default='borderline-1'
+        The type of SMOTE algorithm to use one of the following options:
+        ``'borderline-1'``, ``'borderline-2'``.
+
+    See Also
+    --------
+    SMOTE : Over-sample using SMOTE.
+
+    SMOTENC : Over-sample using SMOTE for continuous and categorical features.
+
+    SVMSMOTE : Over-sample using SVM-SMOTE variant.
+
+    ADASYN : Over-sample using ADASYN.
+
+    KMeansSMOTE : Over-sample applying a clustering before to oversample using
+        SMOTE.
+
+    Notes
+    -----
+    See the original papers: [2]_ for more details.
+
+    Supports multi-class resampling. A one-vs.-rest scheme is used as
+    originally proposed in [1]_.
+
+    References
+    ----------
+    .. [1] N. V. Chawla, K. W. Bowyer, L. O.Hall, W. P. Kegelmeyer, "SMOTE:
+       synthetic minority over-sampling technique," Journal of artificial
+       intelligence research, 321-357, 2002.
+
+    .. [2] H. Han, W. Wen-Yuan, M. Bing-Huan, "Borderline-SMOTE: a new
+       over-sampling method in imbalanced data sets learning," Advances in
+       intelligent computing, 878-887, 2005.
+
+    Examples
+    --------
+    >>> from collections import Counter
+    >>> from sklearn.datasets import make_classification
+    >>> from imblearn.over_sampling import \
+BorderlineSMOTE # doctest: +NORMALIZE_WHITESPACE
+    >>> X, y = make_classification(n_classes=2, class_sep=2,
+    ... weights=[0.1, 0.9], n_informative=3, n_redundant=1, flip_y=0,
+    ... n_features=20, n_clusters_per_class=1, n_samples=1000, random_state=10)
+    >>> print('Original dataset shape %s' % Counter(y))
+    Original dataset shape Counter({{1: 900, 0: 100}})
+    >>> sm = BorderlineSMOTE(random_state=42)
+    >>> X_res, y_res = sm.fit_resample(X, y)
+    >>> print('Resampled dataset shape %s' % Counter(y_res))
+    Resampled dataset shape Counter({{0: 900, 1: 900}})
+    """
+
+    @_deprecate_positional_args
+    def __init__(
+        self,
+        *,
+        sampling_strategy="auto",
+        random_state=None,
+        k_neighbors=5,
+        n_jobs=None,
+        m_neighbors=10,
+        kind="borderline-1",
+    ):
+        super().__init__(
+            sampling_strategy=sampling_strategy,
+            random_state=random_state,
+            k_neighbors=k_neighbors,
+            n_jobs=n_jobs,
+        )
+        self.m_neighbors = m_neighbors
+        self.kind = kind
+
+    def _validate_estimator(self):
+        super()._validate_estimator()
+        self.nn_m_ = check_neighbors_object(
+            "m_neighbors", self.m_neighbors, additional_neighbor=1
+        )
+        self.nn_m_.set_params(**{"n_jobs": self.n_jobs})
+        if self.kind not in ("borderline-1", "borderline-2"):
+            raise ValueError(
+                f'The possible "kind" of algorithm are '
+                f'"borderline-1" and "borderline-2".'
+                f"Got {self.kind} instead."
+            )
+
+    def _fit_resample(self, X, y):
+        self._validate_estimator()
+
+        X_resampled = X.copy()
+        y_resampled = y.copy()
+
+        for class_sample, n_samples in self.sampling_strategy_.items():
+            if n_samples == 0:
+                continue
+            target_class_indices = np.flatnonzero(y == class_sample)
+            X_class = _safe_indexing(X, target_class_indices)
+
+            self.nn_m_.fit(X)
+            danger_index = self._in_danger_noise(
+                self.nn_m_, X_class, class_sample, y, kind="danger"
+            )
+            if not any(danger_index):
+                continue
+
+            self.nn_k_.fit(X_class)
+            nns = self.nn_k_.kneighbors(
+                _safe_indexing(X_class, danger_index), return_distance=False
+            )[:, 1:]
+
+            # divergence between borderline-1 and borderline-2
+            if self.kind == "borderline-1":
+                # Create synthetic samples for borderline points.
+                X_new, y_new = self._make_samples(
+                    _safe_indexing(X_class, danger_index),
+                    y.dtype,
+                    class_sample,
+                    X_class,
+                    nns,
+                    n_samples,
+                )
+                if sparse.issparse(X_new):
+                    X_resampled = sparse.vstack([X_resampled, X_new])
+                else:
+                    X_resampled = np.vstack((X_resampled, X_new))
+                y_resampled = np.hstack((y_resampled, y_new))
+
+            elif self.kind == "borderline-2":
+                random_state = check_random_state(self.random_state)
+                fractions = random_state.beta(10, 10)
+
+                # only minority
+                X_new_1, y_new_1 = self._make_samples(
+                    _safe_indexing(X_class, danger_index),
+                    y.dtype,
+                    class_sample,
+                    X_class,
+                    nns,
+                    int(fractions * (n_samples + 1)),
+                    step_size=1.0,
+                )
+
+                # we use a one-vs-rest policy to handle the multiclass in which
+                # new samples will be created considering not only the majority
+                # class but all over classes.
+                X_new_2, y_new_2 = self._make_samples(
+                    _safe_indexing(X_class, danger_index),
+                    y.dtype,
+                    class_sample,
+                    _safe_indexing(X, np.flatnonzero(y != class_sample)),
+                    nns,
+                    int((1 - fractions) * n_samples),
+                    step_size=0.5,
+                )
+
+                if sparse.issparse(X_resampled):
+                    X_resampled = sparse.vstack([X_resampled, X_new_1, X_new_2])
+                else:
+                    X_resampled = np.vstack((X_resampled, X_new_1, X_new_2))
+                y_resampled = np.hstack((y_resampled, y_new_1, y_new_2))
+
+        return X_resampled, y_resampled
+
+
+@Substitution(
+    sampling_strategy=BaseOverSampler._sampling_strategy_docstring,
+    n_jobs=_n_jobs_docstring,
+    random_state=_random_state_docstring,
+)
+class SVMSMOTE(BaseSMOTE):
+    """Over-sampling using SVM-SMOTE.
+
+    Variant of SMOTE algorithm which use an SVM algorithm to detect sample to
+    use for generating new synthetic samples as proposed in [2]_.
+
+    Read more in the :ref:`User Guide <smote_adasyn>`.
+
+    Parameters
+    ----------
+    {sampling_strategy}
+
+    {random_state}
+
+    k_neighbors : int or object, default=5
+        If ``int``, number of nearest neighbours to used to construct synthetic
+        samples.  If object, an estimator that inherits from
+        :class:`~sklearn.neighbors.base.KNeighborsMixin` that will be used to
+        find the k_neighbors.
+
+    {n_jobs}
+
+    m_neighbors : int or object, default=10
+        If int, number of nearest neighbours to use to determine if a minority
+        sample is in danger. If object, an estimator that inherits from
+        :class:`~sklearn.neighbors.base.KNeighborsMixin` that will be used to
+        find the m_neighbors.
+
+    svm_estimator : estimator object, default=SVC()
+        A parametrized :class:`~sklearn.svm.SVC` classifier can be passed.
+
+    out_step : float, default=0.5
+        Step size when extrapolating.
+
+    See Also
+    --------
+    SMOTE : Over-sample using SMOTE.
+
+    SMOTENC : Over-sample using SMOTE for continuous and categorical features.
+
+    SMOTEN : Over-sample using the SMOTE variant specifically for categorical
+        features only.
+
+    BorderlineSMOTE : Over-sample using Borderline-SMOTE.
+
+    ADASYN : Over-sample using ADASYN.
+
+    KMeansSMOTE : Over-sample applying a clustering before to oversample using
+        SMOTE.
+
+    Notes
+    -----
+    See the original papers: [2]_ for more details.
+
+    Supports multi-class resampling. A one-vs.-rest scheme is used as
+    originally proposed in [1]_.
+
+    References
+    ----------
+    .. [1] N. V. Chawla, K. W. Bowyer, L. O.Hall, W. P. Kegelmeyer, "SMOTE:
+       synthetic minority over-sampling technique," Journal of artificial
+       intelligence research, 321-357, 2002.
+
+    .. [2] H. M. Nguyen, E. W. Cooper, K. Kamei, "Borderline over-sampling for
+       imbalanced data classification," International Journal of Knowledge
+       Engineering and Soft Data Paradigms, 3(1), pp.4-21, 2009.
+
+    Examples
+    --------
+    >>> from collections import Counter
+    >>> from sklearn.datasets import make_classification
+    >>> from imblearn.over_sampling import \
+SVMSMOTE # doctest: +NORMALIZE_WHITESPACE
+    >>> X, y = make_classification(n_classes=2, class_sep=2,
+    ... weights=[0.1, 0.9], n_informative=3, n_redundant=1, flip_y=0,
+    ... n_features=20, n_clusters_per_class=1, n_samples=1000, random_state=10)
+    >>> print('Original dataset shape %s' % Counter(y))
+    Original dataset shape Counter({{1: 900, 0: 100}})
+    >>> sm = SVMSMOTE(random_state=42)
+    >>> X_res, y_res = sm.fit_resample(X, y)
+    >>> print('Resampled dataset shape %s' % Counter(y_res))
+    Resampled dataset shape Counter({{0: 900, 1: 900}})
+    """
+
+    @_deprecate_positional_args
+    def __init__(
+        self,
+        *,
+        sampling_strategy="auto",
+        random_state=None,
+        k_neighbors=5,
+        n_jobs=None,
+        m_neighbors=10,
+        svm_estimator=None,
+        out_step=0.5,
+    ):
+        super().__init__(
+            sampling_strategy=sampling_strategy,
+            random_state=random_state,
+            k_neighbors=k_neighbors,
+            n_jobs=n_jobs,
+        )
+        self.m_neighbors = m_neighbors
+        self.svm_estimator = svm_estimator
+        self.out_step = out_step
+
+    def _validate_estimator(self):
+        super()._validate_estimator()
+        self.nn_m_ = check_neighbors_object(
+            "m_neighbors", self.m_neighbors, additional_neighbor=1
+        )
+        self.nn_m_.set_params(**{"n_jobs": self.n_jobs})
+
+        if self.svm_estimator is None:
+            self.svm_estimator_ = SVC(gamma="scale", random_state=self.random_state)
+        elif isinstance(self.svm_estimator, SVC):
+            self.svm_estimator_ = clone(self.svm_estimator)
+        else:
+            raise_isinstance_error("svm_estimator", [SVC], self.svm_estimator)
+
+    def _fit_resample(self, X, y):
+        self._validate_estimator()
+        random_state = check_random_state(self.random_state)
+        X_resampled = X.copy()
+        y_resampled = y.copy()
+
+        for class_sample, n_samples in self.sampling_strategy_.items():
+            if n_samples == 0:
+                continue
+            target_class_indices = np.flatnonzero(y == class_sample)
+            X_class = _safe_indexing(X, target_class_indices)
+
+            self.svm_estimator_.fit(X, y)
+            support_index = self.svm_estimator_.support_[
+                y[self.svm_estimator_.support_] == class_sample
+            ]
+            support_vector = _safe_indexing(X, support_index)
+
+            self.nn_m_.fit(X)
+            noise_bool = self._in_danger_noise(
+                self.nn_m_, support_vector, class_sample, y, kind="noise"
+            )
+            support_vector = _safe_indexing(
+                support_vector, np.flatnonzero(np.logical_not(noise_bool))
+            )
+            danger_bool = self._in_danger_noise(
+                self.nn_m_, support_vector, class_sample, y, kind="danger"
+            )
+            safety_bool = np.logical_not(danger_bool)
+
+            self.nn_k_.fit(X_class)
+            fractions = random_state.beta(10, 10)
+            n_generated_samples = int(fractions * (n_samples + 1))
+            if np.count_nonzero(danger_bool) > 0:
+                nns = self.nn_k_.kneighbors(
+                    _safe_indexing(support_vector, np.flatnonzero(danger_bool)),
+                    return_distance=False,
+                )[:, 1:]
+
+                X_new_1, y_new_1 = self._make_samples(
+                    _safe_indexing(support_vector, np.flatnonzero(danger_bool)),
+                    y.dtype,
+                    class_sample,
+                    X_class,
+                    nns,
+                    n_generated_samples,
+                    step_size=1.0,
+                )
+
+            if np.count_nonzero(safety_bool) > 0:
+                nns = self.nn_k_.kneighbors(
+                    _safe_indexing(support_vector, np.flatnonzero(safety_bool)),
+                    return_distance=False,
+                )[:, 1:]
+
+                X_new_2, y_new_2 = self._make_samples(
+                    _safe_indexing(support_vector, np.flatnonzero(safety_bool)),
+                    y.dtype,
+                    class_sample,
+                    X_class,
+                    nns,
+                    n_samples - n_generated_samples,
+                    step_size=-self.out_step,
+                )
+
+            if np.count_nonzero(danger_bool) > 0 and np.count_nonzero(safety_bool) > 0:
+                if sparse.issparse(X_resampled):
+                    X_resampled = sparse.vstack([X_resampled, X_new_1, X_new_2])
+                else:
+                    X_resampled = np.vstack((X_resampled, X_new_1, X_new_2))
+                y_resampled = np.concatenate((y_resampled, y_new_1, y_new_2), axis=0)
+            elif np.count_nonzero(danger_bool) == 0:
+                if sparse.issparse(X_resampled):
+                    X_resampled = sparse.vstack([X_resampled, X_new_2])
+                else:
+                    X_resampled = np.vstack((X_resampled, X_new_2))
+                y_resampled = np.concatenate((y_resampled, y_new_2), axis=0)
+            elif np.count_nonzero(safety_bool) == 0:
+                if sparse.issparse(X_resampled):
+                    X_resampled = sparse.vstack([X_resampled, X_new_1])
+                else:
+                    X_resampled = np.vstack((X_resampled, X_new_1))
+                y_resampled = np.concatenate((y_resampled, y_new_1), axis=0)
+
+        return X_resampled, y_resampled
